Add files via upload
This is a .csv format file containing code quality features. All the projects analyzed are created in java. software used for evaluation of parameters is SonarQube.
the regressive testing is done to attain the values. technical debt, however, is in percentage which is converted into a boolean format.
this dataset can be used to predict technical debt using classification models.